archive/briefer-cli/briefer/requirements.txt
# CLI
click>=8.1

# Database
duckdb>=1.0

# HTTP
httpx>=0.27

# AI
anthropic>=0.34.0

# Terminal display
rich>=13.0

# Data / time
python-dateutil>=2.9.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.1
